reorganized components, css modules
