+++ pages/newsletter/index.js
@@ -1,4 +1,4 @@
-import NewsletterCardGallery from "@/components/NewsCard";
+import NewsletterCardGallery from "@/components/newsletter/NewsCard";
 import NewsHeaderCarousel from "./headerCarousel";
 
 export default function Art() {
