+++ styles/art/YearCarousel.module.css
